Added mpi support to oxDNA. It is a very primitive support: the same simulation can be run on N different threads. The output files for each simulations are prefixed with mpi_X_, where X is the rank number of the process. This is very experimental, but you can play with it. Now the compilation process creates another executable, oxDNA_mpi, which should be run as follows:
mpirun -np X /path/to/oxDNA/build/bin/oxDNA_mpi input

where X is the number of processes you would like to run. Note that I have also changed the cmake file, which now should link with mpi and thus generate the oxDNA_mpi executable only if the mpi binaries are found. I still didn't try what happens on a machine not equipped with mpi. In principle the executable should not be created. I will try to test the compilation on a system without mpi.