1) Updated documentation (and removed as many doxygen warnings as I could) 2) Updated the README file 3) Removed a couple of warnings from the TEPInteraction class 4) Updated the coding_guidelines.txt file with a few new lines. 5) Updated a few private things
TODO: using plugins for CUDA interactions is nightmarish at the moment, we have to something. As of now, one needs to create entry points with custom names AND add the used prefix in the input file via the key 'plugin_interaction_entry_points'. I had to also change how the plugin manager searches files for the entry points, so that it would first use the custom entry point prefixes and then the default ones. To summarise, the plugin manager is a bit messy, we need to polish it up a bit.