MERGING OF THE BOX BRANCH
Please report any bugs you find: although I have tested the code, a more extensive use is definitely needed. I'm also interested in performances: from my (limited) tests, it looks like the code is either as fast as before or ever-so-slightly slower (5%).

This is a very substantial revision that adds support for non-cubic boxes. It also changes moved the responsability of computing particles' potential neighbours and neighbours from BaseInteraction to BaseList. I have changed all existing interactions and observables to comply with the new policty. However, this change may have disrupted private stuff you didn't commit in the repository. Check BaseList and BaseInteraction classes for the new
method signatures you should use in the future. Note that some of the functions in BaseInteraction that have not been removed are only wrappers around BaseList's method and may be removed in future versions. Note that non-cubic boxes are not supported by CUDA simulations yet.

I have also changed the cmake's "test" target to "test_quick" in order to comply with the new (3.x) cmake policies that didn't like the name.