Implementation of Checkpoints
(optional) Checkpoints have been implemented, although not thoroughly tested, so that the code can be restarted and follow the same trajectory. See the Doxygen documentation for SimBackend on how to use them. This cannot work on CUDA due to the operation order not being predictable.

Other than introducing this new feature, the code should behave exactly the same from the users perspective.

Most of the changes are due to a change in signature of SimBackend::init(), which was overdue, and a change in signature in BaseList::global_update().