Added the UnstackedList observable and the dummy_topology.py script.
The UnstackedList observable prints a list of all the nucleotides that are NOT
stacked with the nucleotide with the successive index. 

The dummy_topology.py script produces a topology file for a double strand of N base pairs
where hydrogen bonding is allowed only between bases that face each other in the native
conformation. This can be used to study average duplexes where mismatches are forbidden.