Improvements in the function getParticlesFromString, that parses a string containing particle identifiers. It is currently used only by the forces "string" and "repulsion_plane_moving", but eventually should be used by all observables and forces.
Up until this moment it could parse a comma separated list of particle identifiers, that could be any of the following:
1) particle indices 
2) particle indices intervals separated by a dash (so that "particle = 0, 2, 3, 4" is equivalent to "particle = 0, 2-4" if particles 2, 3 and 4 are neighbours along the n5 direction). 
3) -1, that expands to all the particles in the system.

Now it also understands the identifiers 
4) "last" (that automatically maps to the bead with the highest index)
5) "all" (equivalent to -1, but possibly more explicative)

Also it has now been made more robust and hopefully more communicative when it finds something that doesn't work. 
It's been tested and it seems to be working fine, but please use with a grain of salt.