Updated the TEP interaction and the Utils file in the nBlock branch
Now the TEP interaction in the nBlock branch should behave like the one in the
main branch, with the difference that the last bead of the TEP chain have a 
non-unitary mass.

The mass is set with the key
TEP_last_particle_mass = <mass>
in the input file (where <mass> is a number).