Added a new feature to TEPInteraction: randomly fluctuating twist
This is implemented as a new argument TEP_choose_direction_time, which can only be used without TEP_max_twisting_time.
The argument specifies a time at which both ends of the strand will randomly choose whether to change the direction of rotation,
i.e. from clockwise to counter-clockwise or viceversa.

If the the beads change direction of rotation, they do so at the same time.

I.e. if you want to use this and you're happy with how your beads spin when you are using the TEP_max_twisting_time argument
to change the linking number of the system, using the TEP_choose_direction_time instead should give you a system whose 
linking number performs an unbiased random walk, and the beads will never be spinning in the same direction. This guarantees
that the linking number is never constant.

It compiles, it runs, and seems to work, but of course caution should be used.