Added a sequence dependence file that reproduces the average parametrisation of oxDNA2.
Using sequence dependence with this file is equivalent to not using it at all (i.e. using the average sequence parametrisation of oxDNA2).
If you want to study a system that behaves like the average model almost in every aspect, with some modifications (i.e. no hydrogen bonding, or AA stacking set to some value etc.), you can use this file as a starting point.