Fixed a problem that would arise when sformat (which is internally used by oxDNAException) would generate strings longer than 500 characters.
As stated in the docs, vsnprintf internally retrieves arguments from the list identified by its last argument (arg) as if va_arg was used on it, and thus the state of arg is likely to be altered by the call. Therefore, before calling it it is necessary to make a copy of arg, look whether there is enough allocated space for the output string and then return the given string. If the output string is not long enough, more space needs to be allocated and vsnprintf called once more.