newrelax: ConfigInfo is now a singleton, thus accessible from any object.
ConfigInfo (previously a class defined in the Observables/BaseObservable.h) was an ancillary class initially devised as a simple way of passing information from the backend to the observables. It has been since abused to pass around some quantities and pointers throughout the code. Since some classes of objects (observables, for instance) have access to it but some others (like boxes) don't, it's probably better to make it accessible from everywhere. I thus decided to move it to the Utilities folder and transform it into a singleton. This change required to change many files, even though the changes themselves are quite small. I have run "make test_run" and "make test_quick" and both didn't fail. I have also tested MC2 by hand and it works. The chance also affects the CUDA backend, which I could not test but will do in the near future.