newrelax: implemented the DPD thermostat
The DPD thermostat is a Langevin-like thermostat that acts on the relative velocities of each pair of particles. It can be very useful in out-of-equilibrium scenarios (such as simulations under shear flow). As far as I know, the DPD thermostat has never been extended to rigid bodies, which means that, for now, in oxDNA this thermostat affects only the translational degrees of freedom.