Substantial commit *only* for MC_CPUBackend2 users (if you don't know what it is you aren't).
Now moves have two more keys (compute_energy_before and restrict_to_type), where
the names should be self-explanatory.  compute_energy_before (default: true) can
be used to turn off the energy computation before the move. This is useful to
reduce the calls to particle_energy by a factor 2 ONLY IF the interaction is
hard and purely repulsive (such as hard spheres).  restrict_to_type (default:
-1, i.e. no effect) will restrict that particular move only to particles with
the given particle type. It is up to the user to make the particle type
consistent with that set up when reading the topology by the interaction class.
This is useful if one deals with mixtures. For example, one may want to use a
large step for fast particles and a small step for slow particles.

There is also one extra method in SimBackend that allows to print some
information after the equilibration steps. This is useful for example to print
out the size of the moves after they have been adjusted.

The output of the moves has also been made much more verbose to better
understand what the code is doing.