Generalise the mechanism through which CPU and GPU data are synced
Before this commit, the CUDA backends copied the data from the GPU to
the CPU and back through their own methods. Now this logic has been
incapsulated in SimBackend's methods that are now used when observables
are printed, when fix diffusion is called and when SimManager's and
OxpyManager's run methods are invoked.