Update the way code versioning is handled
Now the executables print both the latest release (which is the current
git tag) and the current commit (**if** the user compiling the code has
cloned the repo with git).

Note that from now on the current git tag and the content of the RELEASE
macro defined in `src/defs.` will have to match or the pre-commit hook
script will prevent the user from committing.