Add data structures to handle "molecules"
* Molecules are groups of particles connected by bonds that are not
broken throughout the simulation. For DNA and RNA, these are the
strands. SimBackend has a vector of molecule, and ConfigInfo a pointer
to it that can be accessed by any other class. As of now, in the core
classes only fix_diffusion uses these new data structures.
* Add a volume move that acts on the centres of mass of the molecules.
It does not work for now.