Add a rudimentary subscriber/notifier mechanism
For now the mechanism can be (and is) used to manage changes to the
temperature. The temperature can be changed by using `ConfigInfo`'s
`update_temperature` method. Instances that want to be notified by the
change has to subscribe to the "T_updated" event by using `ConfigInfo`'s
`subscribe` method.