Fix the way the analysis backend works
The way DNAnalysis is used remains unchanged, but under the hood now the
method that reads the next configuration also updates the lists (while
before the two operations were done one after the other by the analyse()
method).

This makes it easier to do the analysis of a trajectory from Python like
this:

    while backend.read_next_configuration():
        print(backend.particles[0].pos)

Before this commit the above while would start from the *second*
configuration rather then the first.