Observables can now update their data structures while the sim runs
The frequency with which their `update_data` method is invoked is
controlled by the `update_every` key and need not be the same as their
associated output's print_every. In this way some of the analysis
(autocorrelations, averages) can be performed at a high frequency, while
the result of these operations can be printed with a lower frequency in
order to not decrease performance or increase disk usage.