Fix a bug whereby a DNA force was sometimes incorrectly evaluated on GPU
On both CPU and GPU the derivative of the f4 function was evaluated and
then divided by sin(t), where t is the function's argument. The
numerator is (t - t0), where in some cases t0 is 0. In these cases one
can use the fact that t / sin(t) -> 1 when t -> 0 to avoid the numerical
instability caused by 1 / sin(t) when t -> 0. However, this is incorrect
when t0 != 0. This commit fixes this issue on GPU only.