Add the possibility of disabling JSON support at compile time
This commit adds a JSON_ENABLED option to the pre-compilation (cmake)
stage. By default JSON_ENABLED is set to On. If set to Off, JSON support
is disabled and the nlhomann library is not used during compilation.
This should help to compile the code if some old compilers are used
(see #59)