Fix a bug with the DNA*_nomesh interactions
On CPU, two nucleotides which were perfectly aligned to form hydrogen bonds (as in generated duplex configurations) would result in wrong energies in the first simulation step. This was due to using the acos function with no guards (i.e. code that checks whether the argument is out of boundary)