Add a pyproject.toml and slightly simplify how oxpy compilation is carried out (#180)
Use pyproject.toml instead of the custom setup.py we had before. Note that it is still possible to compile&install the package either with `pip install .` or by compiling with `-DPython=On` and then install the package with `make install`.

A rather comprehensive explanation for how this is done (or at least, how I managed to make it work) can be found in this "blog" post: https://www.roma1.infn.it/~rovigatl/posts/pybind_cmake/

---------

Co-authored-by: Erik Poppleton <rkpoppleton@gmail.com>